Don't show delete all alert when no image is selected

+++ app/src/main/java/swati4star/createpdf/fragment/ViewFilesFragment.java
@@ -105,7 +105,9 @@ public class ViewFilesFragment extends Fragment implements SwipeRefreshLayout.On
                 displaySortDialog();
                 break;
             case R.id.item_delete:
-                deleteFiles();
+                if(mViewFilesAdapter.areItemsForDeleteSelected()) {
+                    deleteFiles();
+                }
                 break;
             default:
                 break;
